refactor: Remove src/auth/ directory — relocate JWKS + models to src/api/auth
The old src/auth/ package (OAuth providers, sessions, tokens, 2FA,
verification) was fully migrated to user-service in earlier waves. Only
the JWKS validation and shared auth models remained. This commit:

- Consolidates jwks.py and models.py into src/api/auth.py
- Deletes src/auth/ directory entirely (3 files, 136 lines)
- Updates all import sites in src/ and tests/ (18 files)
- All 496 tests pass, lint and typecheck clean

Closes #758
