+++ tests/test_auth/conftest.py
@@ -53,7 +53,7 @@ def _patch_settings(test_settings: Settings) -> Iterator[None]:
     """Patch get_settings at every import site so all auth code uses test settings."""
     targets = [
         "src.config.get_settings",
-        "src.auth.jwks.get_settings",
+        "src.api.auth.get_settings",
     ]
     patches = [patch(t, return_value=test_settings) for t in targets]
     for p in patches:
